requirements.txt
google-generativeai
pandas
pillow
